// client/model/getManyEndpointGroupResponseDto.ts
/**
 * Draas
 * Draas OAS Documentation
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { EndpointGroup } from './endpointGroup';


export interface GetManyEndpointGroupResponseDto { 
    totalPages: number;
    count: number;
    total: number;
    page: number;
    pageCount: number;
    data: Array<EndpointGroup>;
}